Adjust PullRequests before writing to file
Incremental loading did not work very well because the repo's
stats would keep on changing, which would cause new and existing files
to have diffs like this:
```
% diff ~/Desktop/18458-new.json ~/Desktop/18458.json
182,183c182,183
<       "updated_at": "2021-12-04T13:14:18Z",
<       "pushed_at": "2021-12-04T13:14:15Z",
---
>       "updated_at": "2021-11-23T10:20:20Z",
>       "pushed_at": "2021-11-23T13:13:29Z",
189c189
<       "size": 158605,
---
>       "size": 152001,
202c202
<       "open_issues_count": 57,
---
>       "open_issues_count": 72,
```

So in essence, the fact that the Repository object would change,
all incremental loads would turn into full loads due to this.

To fix this, adjust the PullRequest object by zeroing out the
offending fields. This has the risk that some other field could
keep on changing, but it's the least expensive way of doing this.