Move repo total aggregation to backend, compute Comments / Request
Comments / Request is a simple ratio of the number of comments the
engineer authored to others for every review request they received.
A very low number could mean they're not very engaged, or everything
their colleagues submit is of top quality. In other words - it does
not indicate anything in isolation, but the value is interesting
relative to rest of the team.