Compute partial daily results cache
This additional, intermediate cache layer produces files
per date and per user id. Once this cache is computed, it
allows for much faster returning of data regardless of which
date period is used, or which users are included in the result.
"Much faster" can mean ~5sec returning of results instead of
waiting for several minutes.

Mutable merging is used to avoid performance impact when
merging per-user results together; profiling showed that
bulk of the time was spent creating copies of objects on
merge ops.

The response-level cache is still kept.